% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{Clark2dt.log}
\alias{Clark2dt.log}
\title{Clark2dt.log}
\usage{
Clark2dt.log(x, par)
}
\arguments{
\item{x}{Numeric vector of distances the the nearest seed source.}

\item{par}{Numeric vector with three elements representing the log-transformed parameters \eqn{u} and \eqn{p} and the scaling \eqn{N}.}
}
\value{
Numeric vector of function values multiplied by \eqn{N}.
}
\description{
Clark2dt computes the value of the dispersal function from Clark et al. (1999) multiplied by \eqn{N}.
}
\details{
Mixture of Gaussian nuclei that produces tails that are not quite as long. Maximum at seed tree itself and cannot become 0 at x = 0.
}
