% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{Clark2dt}
\alias{Clark2dt}
\title{Clark2DT-Function}
\usage{
Clark2dt(x, par)
}
\arguments{
\item{x}{represents the distance to the nearest seed source. Must be numeric}

\item{par}{parameters U, P, N estimated within the Clark2DT function}
}
\value{
what is returned by the function?
}
\description{
Clark2DT-Function
}
\details{
Mixture of Gaussian nuclei that produces tails that are not quite as long. Maximum at seed tree itself and cannot become 0 at x = 0.
}
