% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distmap.R
\name{Distmap}
\alias{Distmap}
\title{Distance map to nearest seed source}
\usage{
Distmap(fe_raster, treespecies)
}
\arguments{
\item{fe_raster}{Remote sensing raster dataset with tree species classification of specific tree species and tree species groups.}

\item{treespecies}{Represents the numerical value by which the tree species of interest was encoded in the raster dataset.}
}
\value{
The distance is given in m.
}
\description{
Creation of a distance map for the study area. The distance to
the nearest seed source is calculated for every raster cell.
}
\details{
A SpatRaster is created with the same resolution as the input raster, defined by the study area.
}
\examples{
## Raster data set
rr <- terra::rast(
 matrix(sample(0:10, 20 * 20, replace = TRUE),
        nrow = 20, ncol = 20))

## compute distance for prediction area
distance <- Distmap(fe_raster = rr, treespecies = "10")
terra::plot(distance)

}
