% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{geometric.log}
\alias{geometric.log}
\title{Geometric.log-Function}
\usage{
geometric.log(x, par)
}
\arguments{
\item{x}{represents the distance to the nearest seed source. Must be numeric.}

\item{par}{parameters a, b, N estimated within the geometric.log function with a and b representing the scale and the shape parameter of the function, respectively.}
}
\description{
Geometric.log-Function
}
\details{
The function will behave quite differently from the exponential and Weibull distributions. They show a fat tail, whatever the value of the shape parameter (b), and the distributions become increasingly fat-tailed ?as b declines toward ‘1’.
}
