% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{lognormal.log}
\alias{lognormal.log}
\title{lognormal.log}
\usage{
lognormal.log(x, par)
}
\arguments{
\item{x}{represents the distance to the nearest seed source. Must be
numeric.}

\item{par}{Numeric vector with three elements representing log-transformed
scale and shape parameters, given by the mean \eqn{a} and standard
deviation \eqn{\sigma} of the underlying normal distribution, and the
scaling \eqn{N}.}
}
\description{
lognormal.log
}
\details{
The spatial dispersal density, representing the probability
density function of the distance of a seed from its source divided by
\eqn{2\pi x}, is here given by
\deqn{f(x) = \frac{1}{2\pi x^2 \sqrt{2\pi\sigma^2}} 
  e^{-\frac{1}{2\sigma^2}(\log(x/a))^2},}
see Greene and Johnson (1989), Stoyan and Wagner (2001). Thus, the
distance is assumed to have the log-normal distribution such that the
log-distance has a normal distribution with mean \eqn{a} and variance
\eqn{\sigma^2}. Note that \eqn{\log f(x)} is a quadratic function of
\eqn{\log x} with a maximum at \eqn{\log a-2\sigma^2}.
}
\references{
Greene, D.F., Johnson, E.A., 1989. A model of wind dispersal of winged or
plumed seeds. \emph{Ecology} \strong{70}(2), 339–347.
\doi{10.2307/1937538}

Stoyan, D., Wagner, S., 2001. Estimating the fruit dispersion of
anemochorous forest trees. \emph{Ecol. Modell.} \strong{145}, 35–47.
\doi{10.1016/S0304-3800(01)00385-4}
}
