% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{lognormal}
\alias{lognormal}
\title{lognormal-Function}
\usage{
lognormal(x, par)
}
\arguments{
\item{x}{represents the distance to the nearest seed source. Must be numeric}

\item{par}{parameters a, σ, N estimated within the lognormal function}
}
\description{
lognormal-Function
}
\details{
The equation means log η_i is a quadratic function of log x_i with a maximum at log a-2σ^2.
}
