% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optN.R
\name{optN}
\alias{optN}
\title{optN}
\usage{
optN(par, nframe, tol, ...)
}
\arguments{
\item{par}{are parameters to be estimated}
}
\description{
For given a,σ the global minimum in N can always be found with any desired precision, usually in a small number of steps, by successively shrinking an interval. We realize this as an inner, nested minimization in the function optN.
}
