% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{power}
\alias{power}
\title{Power-Law Dispersal Kernels}
\usage{
power(x, par)
}
\arguments{
\item{x}{represents the distance to the nearest seed source. Must be
numeric.}

\item{par}{Numeric vector with three elements representing the
log-transformed parameters \eqn{a} and \eqn{b} and the scaling \eqn{N}.}
}
\value{
Numeric vector of function values multiplied by \eqn{N}.
}
\description{
\code{power} computes the value of the dispersal function from (WHERE?)
multiplied by \eqn{N}.
}
\details{
The dispersal kernel, i.e. spatial probability density
of the location of a seed relative to its source, is here given by
\deqn{k(x)={\Gamma (d/2) \over 2\pi ^{d/2}a^{d}\Beta(d,b-d)}
  (1+{\left\|{x}\right\| \over a})^{-b},}
which corresponds to a probability density of the distance given by
\deqn{p(x)={1 \over a^{d}\Beta(d,b-d)}r^{d-1}(1+{r \over a})^{-b},}
where \eqn{d} is the spatial dimension and \eqn{\left\|{\,}\right\|}
denotes the Euclidean norm; see Nathan et al. (2012) for the planar case
(DOUBLE CHECK EQUATIONS).

The spatial dispersal density, representing the probability
density function, divided by \eqn{2\pi x}, of the distance of a seed from
its source, is here given by
\deqn{f(x) = \frac{(b-2)(b-1)}{2\pi a^2} (1+\frac{x}{a})^{-b},}
see Nathan et al. (2012).
(CHANGE THE FOLLOWING, SHOULD BE OUR OWN CHARACTERIZATION:) Austerlitz
et al. (2004) characterize it as follows: The geometric and 2dt families
“will behave quite differently from the exponential and Weibull
distributions. They show a fat tail, whatever the value of the shape
parameter (\eqn{b}), and the distributions become increasingly fat-tailed
as \eqn{b} declines toward ‘1’.”
}
\references{
Nathan, R., Klein, E., Robledo‐Arnuncio, J.J., Revilla, E. (2012).
Dispersal kernels: review, in Clobert, J., Baguette, M., Benton, T.G.,
Bullock, J.M. (eds.), \emph{Dispersal ecology and evolution}, 186–210.
\doi{10.1093/acprof:oso/9780199608898.003.0015}

Austerlitz, F., Dick, C.W., Dutech, C., Klein, E.K., Oddou-Muratorio, S.,
Smouse, P.E., Sork, V.L. (2004). Using genetic markers to estimate the
pollen dispersal curve. \emph{Molecular Ecology} \strong{13}, 937–954.
\doi{10.1111/j.1365-294X.2004.02100.x}
}
