% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optN.R
\name{quax}
\alias{quax}
\title{SN}
\usage{
quax(...)
}
\arguments{
\item{...}{Further arguments passed to \code{optim}.}

\item{x}{Numeric vector giving the distance to the nearest seed source for the inventory plot.}

\item{y}{Observed regeneration density of the inventory plot.}

\item{tau}{Numeric between 0 and 1. Specifies the quantile used for the quantile regression. {0;1}}

\item{fun}{Function assumed for the quantile regression of the regeneration potential. Values allowed are: "exponential.power.log", "weibull.log", "gamma", "Clark2dt.log", "lognormal", "geometric.log". The default is to fit an lognormal model.}

\item{weights}{Numeric vector of weights of the observations in the estimation procedure. Default is 0.}

\item{par}{Numeric vector of initial values for the parameters to be optimized over, exluding the first parameter \code{N}.}

\item{formula}{A formula of the form \code{y ~ x}.}

\item{data, subset, weights, na.action, offset}{For the formula interface: Further arguments passed to \code{model.frame} (along with \code{weights}).}

\item{tol}{The desired accuracy for the inner optimization, see \code{optimize}.}
}
\value{
The estimated function, including an attribute \code{o} containing the results of \code{optim}.
}
\description{
The dispersal function for estimating the potential regeneration density in the 2-dimensional space.
}
\details{
The function return a list including the estimated parameters for the quantile regression for the specific distribution function.
}
