% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optN.R
\name{SN}
\alias{SN}
\title{SN}
\usage{
SN(x, y, tau, fun, par, Nmax = 1e+06)
}
\arguments{
\item{x}{represents the distance to the nearest seed source. Must be numeric}

\item{y}{response: regeneration density in corresponding unit (per ha)}

\item{tau}{determines the quantile}

\item{fun}{dispersal function, which is used}

\item{par}{are parameters to be estimated}
}
\description{
SN
}
