% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quax.R
\name{summary.quax}
\alias{summary.quax}
\title{summary.quax}
\usage{
\method{summary}{quax}(f)
}
\value{
The quality criterion is the value that is minimised in the quantile regression. It represents the weighted sum of absolute residuals.
}
\description{
The function for printing the summary of the quantile regression.
}
\details{
The function return a list including the estimated parameters for the quantile regression for the specific distribution function.#'
The estimated function, including an attribute \code{o} containing the results of \code{optim}.
}
\examples{
#create dataframe
simulated.data <- data.frame(distance = rlnorm(200, meanlog = 5, sdlog = 1), density = rep(0:10,200))

# run quax function
f1 <- quax(x = simulated.data$distance, y = simulated.data$density, tau = 0.9, fun = k_lognormal)

# run summary.quax
summary.quax(f1)

}
