% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{Weibull.log}
\alias{Weibull.log}
\title{Weibull.log-Function}
\usage{
Weibull.log(x, par)
}
\arguments{
\item{x}{represents the distance to the nearest seed source. Must be numeric.}

\item{par}{parameters a, b, N estimated within the Weibull.log function with a and b representing the scale and the shape parameter of the function, respectively.}
}
\description{
Weibull.log-Function
}
\details{
The distribution is fat-tailed when b ≤ 1 and thin-tailed otherwise. As for the exponential power function, when b = 2, the Weibull degenerates to the normal distribution, but when b = 1, it does not degenerate to the exponential distribution.
}
