% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{Clark2dt.log}
\alias{Clark2dt.log}
\title{Dispersal Densities From 2-Dimensional t Distribution}
\usage{
Clark2dt.log(x, par)
}
\arguments{
\item{x}{Numeric vector of distances to the nearest seed source.}

\item{par}{Numeric vector with three elements representing the log-
transformed parameters \eqn{u} and \eqn{p} and the scaling \eqn{N}.}
}
\value{
Numeric vector of function values multiplied by \eqn{N}.
}
\description{
\code{Clark2dt.log} computes the value of the dispersal function from Clark et
al. (1999) multiplied by \eqn{N}.
}
\details{
The spatial dispersal density, representing the probability
density function, divided by \eqn{2\pi x}, of the distance of a seed from
its source, is here given by
\deqn{f(x) = \frac{p}{\pi u (1+x^2/u)^{p+1}},}
see Clark et al. (1999), and Austerlitz et al. (2004) with a different
parameterization (\eqn{b=p+1}). This represents a mixture of Gaussian
densities that produces tails that are not quite as long ???longer???. It
has its maximum at the seed tree itself and cannot become \eqn{0} at
\eqn{x=0}.
}
\references{
Clark, J.S., Silman, M., Kern, R., Macklin, E. and HilleRisLambers, J.
(1999). Seed dispersal near and far: patterns across temperate and tropical
forests. \emph{Ecology} \strong{80}, 1475–1494.
\doi{10.1890/0012-9658(1999)080[1475:SDNAFP]2.0.CO;2}

Austerlitz, F., Dick, C.W., Dutech, C., Klein, E.K., Oddou-Muratorio, S.,
Smouse, P.E. and Sork, V.L. (2004). Using genetic markers to estimate the
pollen dispersal curve. \emph{Molecular Ecology} \strong{13}, 937–954.
\doi{https://doi.org/10.1111/j.1365-294X.2004.02100.x}
}
