% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{Clark2dt}
\alias{Clark2dt}
\title{Dispersal Kernels From Spatial t Distribution}
\usage{
Clark2dt(x, par)
}
\arguments{
\item{x}{Numeric vector of distances to the nearest seed source.}

\item{par}{Numeric vector with three elements representing the
log-transformed parameters \eqn{a} and \eqn{p} and the scaling \eqn{N}.}
}
\value{
Numeric vector of function values of multiplied by \eqn{N}.
}
\description{
\code{Clark2dt} computes the value of the dispersal function from Clark et al.
(1999) multiplied by \eqn{N}.
}
\details{
The dispersal kernel, i.e. spatial probability density
of the location of a seed relative to its source, is here given by
\deqn{k(x)={\Gamma (d/2) \over \pi ^{d/2}a^{d}\Beta (d/2,p)}
  (1+{\left\|{x}\right\|^{2} \over a^{2}})^{-(p+d/2)},}
which corresponds to a probability density of the distance given by
\deqn{p(r)={2 \over a^{d}\Beta (d/2,p)}r^{d-1}
  (1+{r^{2} \over a^{2}})^{-(p+d/2)},}
where \eqn{d} is the spatial dimension, \eqn{\left\|{\,}\right\|}
denotes the Euclidean norm and the normalizing constants involve the
\link[base:beta]{beta} and \link[base:beta]{gamma} functions; see Clark
et al. (1999) and Austerlitz et al. (2004) for the planar case (with
parameterizations \eqn{a=\sqrt{u}} and \eqn{p=b-d/2}, respectively). This
means the squared distance is \eqn{da^{2} \over 2p} times a random
variable having an \link[stats:FDist]{F distribution} with \eqn{d} and
\eqn{2p} degrees of freedom.

This dispersal kernel represents a mixture of Gaussian densities (with
inverse variance following a gamma distribution with shape parameter
\eqn{p}) that exhibits heavier tails. It has its maximum at zero.
}
\references{
Clark, J.S., Silman, M., Kern, R., Macklin, E., HilleRisLambers, J.
(1999). Seed dispersal near and far: patterns across temperate and tropical
forests. \emph{Ecology} \strong{80}, 1475–1494.
\doi{10.1890/0012-9658(1999)080[1475:SDNAFP]2.0.CO;2}

Austerlitz, F., Dick, C.W., Dutech, C., Klein, E.K., Oddou-Muratorio, S.,
Smouse, P.E., Sork, V.L. (2004). Using genetic markers to estimate the
pollen dispersal curve. \emph{Molecular Ecology} \strong{13}, 937–954.
\doi{10.1111/j.1365-294X.2004.02100.x}
}
