% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distmap.R
\name{Distmap}
\alias{Distmap}
\title{Distance map to nearest seed source}
\usage{
Distmap(raster, treespecies)
}
\arguments{
\item{raster}{Remote sensing raster data set with tree species
classification of specific tree species and tree species groups.}

\item{treespecies}{Represents the numerical value by which the tree species
of interest is encoded in the raster data set.}
}
\value{
The distance is given in m.
}
\description{
Creation of a distance map for the study area. The distance to
the nearest seed source is calculated for every raster cell.
}
\details{
A SpatRaster is created with the same resolution as the input raster,
defined by the study area.
}
\examples{
## Create raster data set
 rr <- terra::rast(
 matrix(sample(0:10, 20 * 20, replace = TRUE),
        nrow = 20, ncol = 20))

## Compute distance for study area
distance <- Distmap(raster = rr, treespecies = "10")

## Plot the Distmap
terra::plot(distance)

}
