% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{exponential.power.log}
\alias{exponential.power.log}
\title{exponential.power.log-Function}
\usage{
exponential.power.log(x, par)
}
\arguments{
\item{x}{represents the distance to the nearest seed source. Must be numeric.}

\item{par}{parameters a, b, N estimated within the exponential.power.log function, with a and b representing the scale and the shape parameter of the function, respectively.}
}
\description{
exponential.power.log-Function
}
\details{
The equation means a fat-tailed distributions with the maximum at zero. As a result of its flexible shape, the exponential power distribution has been applied in a number of theoretical studies that address dispersal.
}
