% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_quax.R
\name{predict_quax}
\alias{predict_quax}
\title{predict_quax}
\usage{
predict_quax(distmap, quax)
}
\arguments{
\item{distmap}{An object of the Distmap-Function. A SpatRaster with distances to the nearest seed tree is used for the prediction of the potential regeneration densities.}

\item{quax}{An quax object is used for the prediction. This is a parameterised dispersal function using quantile regression.}
}
\value{
The regeneration density is given in N/ha.
}
\description{
Prediction of the potential regeneration density as a function of the distance to the nearest seed tree.
}
\details{
A SpatRaster is created with the same resolution as the input raster, defined by the study area. The potential regeneration density is calculated and given for each raster cell.
}
