% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optN.R
\name{quax}
\alias{quax}
\title{quax}
\usage{
quax(x, y, tau, fun = lognormal)
}
\arguments{
\item{x}{numeric vector giving the distance to the nearest seed source for the inventory plot.}

\item{y}{Observed regeneration density of the inventory plot.}

\item{tau}{Represents the quantile, which is used for the quantile regression. {0;1}}

\item{fun}{function assumed for the quantile regression of the regeneration potential.Values allowed are: "exponential", "weibull", "gamma", "Clark2dt", "lognormal", "half-normal". The default is to fit an lognormal model.}
}
\value{
The parameter values are given. In the next step the N optimization is done.
}
\description{
The function for estimating the potential regeneration density.
}
\details{
The function return a list including the estimated parameters for the quantile regression for the specific distribution function.
}
