% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_stuff.R
\name{S}
\alias{S}
\title{objective function for minimization}
\usage{
S(N, par, y, tau, w, fun, ...)
}
\arguments{
\item{par}{are parameters to be estimated}

\item{y}{response: regeneration density in corresponding unit (per ha)}

\item{tau}{determines the quantile}

\item{fun}{dispersal function, which is used}

\item{x}{represents the distance to the nearest seed source. Must be numeric}
}
\description{
The parameters are estimated by minimizing a sum of weighted absolute errors, where y_1,...,y_n are the observed regeneration numbers and x_1,...,x_n,η_1,...,η_n are defined as above.
While common regression approaches assume that the average error in the model is zero, the errors are asymmetrically weighted for quantile regression (see Fahrmeir et al., 2013, Ch. 10).
The objective function is defined and the sum of absolute weighted errors is minimized.
\deqn{\sum_{i=1}^n x_i w_i |y_i - \eta_i|}
}
\details{
...
}
