% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{Weibull}
\alias{Weibull}
\title{Dispersal Kernels From Weibull Family}
\usage{
Weibull(x, par)
}
\arguments{
\item{x}{Numeric vector of distances to the nearest seed source.}

\item{par}{Numeric vector with three elements representing the
log-transformed parameters \eqn{a} and \eqn{b} and the scaling \eqn{N}.}
}
\value{
Numeric vector of function values multiplied by \eqn{N}.
}
\description{
\code{Weibull} computes the value of the dispersal function from Tufto et al.
(1997) multiplied by \eqn{N}.
}
\details{
The dispersal kernel, i.e. spatial probability density
of the location of a seed relative to its source, is here given by
\deqn{k(x)={b\Gamma (d/2) \over 2\pi ^{d/2}a^{b}}\left\|{x}\right\|^{b-d}
  e^{-(\left\|{x}\right\|/a)^{b}},}
which corresponds to a probability density of the distance given by
\deqn{p(r)={b \over a^{b}}r^{b-1}e^{-(r/a)^{b}},}
where \eqn{d} is the spatial dimension, \eqn{\left\|{\,}\right\|}
denotes the Euclidean norm and the normalizing constants involve the
\link[base:beta]{gamma} function; see Tufto et al. (1997) for the planar
case. Thus, the distance is assumed to have the
\link[stats:Weibull]{Weibull distribution} with scale parameter \eqn{a}
and shape parameter \eqn{b}. Equivalently, the \eqn{b}th power of the
distance has an exponential distribution with scale parameter \eqn{a}.

The distribution is fat-tailed when \eqn{b<1} and
thin-tailed otherwise (Nathan et al. 2012).
For \eqn{b>1}, the mode of the function is at \eqn{x>1}. In this way, the function approaches the normal distribution.
}
\references{
Tufto, J., Engen, S., Hindar, K. (1997). Stochastic dispersal processes in
plant populations, \emph{Theoretical Population Biology} \strong{52}(1), 16–26.
\doi{10.1006/tpbi.1997.1306}

Austerlitz, F., Dick, C.W., Dutech, C., Klein, E.K., Oddou-Muratorio, S.,
Smouse, P.E., Sork, V.L. (2004). Using genetic markers to estimate the
pollen dispersal curve. \emph{Molecular Ecology} \strong{13}, 937–954.
\doi{10.1111/j.1365-294X.2004.02100.x}

#' Nathan, R., Klein, E., Robledo‐Arnuncio, J.J., Revilla, E. (2012).
Dispersal kernels: review, in Clobert, J., Baguette, M., Benton, T.G.,
Bullock, J.M. (eds.), \emph{Dispersal ecology and evolution}, 186–210.
\doi{10.1093/acprof:oso/9780199608898.003.0015}
}
